+++ docs/CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## v1.1.0.39
 
-- changed to use RSS Atom feed for each version of windows. This feed is kept up to date by microsoft
-- Changed process to determin latest update based on windows 10 build minor number instead of KB number
+- Changed to use RSS Atom feed for each version of Windows. This feed is kept up to date by Microsoft
+- Changed process to determine latest update based on Windows 10 build minor number instead of KB number
 
 ## v1.1.0.38